Avoid ignoring hidden files
Hidden files should generally not be generated by build scripts, so
they do not need to be ignored.  And some hidden files
(e.g. .gitignore) are quite useful to keep under revision control.

If there are specific hidden files that do need to be ignored, they
should be added individually.