Store authority key directly, and move from gpg to gpgv.
We were storing the fingerprint and the signing subkey long key ID,
and then fetching the key from the keyservers and importing it into
the superuser's GnuPG keyring.

It makes more sense to just store the key itself directly, and not
need to worry about potential keyserver problems. This also avoids
tampering with the superuser's keyring.

Furthermore, gpgv is a simpler tool than gpg itself, and is more
narrowly-tailored for signature verification.  So we use the simplest
tool that gets the job done.

Additionally, we explicitly specify both the signature file and the
file that is to be verified, to avoid a common trap where the
signature file itself contains a signed message, leaving the external
file essentially ignored.

Note that the key is stored here in simple base64 format, rather than
as an OpenPGP "ascii-armored" blob, because gpgv prefers its keyrings
in raw binary form.

I considered placing the key in a separate file, but then you need the
script to know where the other file is, and that seemed like more
hassle than it was worth.

If the key itself gets updated in the future, or if you switch keys,
you can update them here, using something like the following to
generate the base64 encoded data:

    gpg --export-options export-minimal \
      --export-filter 'drop-subkey=usage!~s'
      --export B693F33372E965D76D55368616EEA65D03326C9D \
      | base64